backend/requirements.txt
# Private-C privacy intelligence API — production stack
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
google-generativeai>=0.8.0
python-multipart>=0.0.12
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
